Refactored matrix operations
See
https://github.com/advancedresearch/poi/issues/800#issuecomment-72767918
5
